requirements.txt
alembic==1.16.5
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cloudinary==1.44.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pillow==11.3.0
pillow_heif==1.1.1
psycopg2-binary==2.9.11
python-docx==1.2.0
pytz==2025.2
qrcode==8.2
reportlab==4.4.10
six==1.17.0
SQLAlchemy==2.0.46
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
WTForms==3.2.1
zipp==3.23.0
